refactor: reuse existing mock
